+++ changelog.md
@@ -2,6 +2,34 @@
 
 ## 待解决
 
+## [1.0.03] - 2021-07-?
+
+### Added
+
+## [1.0.02] - 2021-07-19
+
+### Added
+
+- Digraph_OSM
+  - download_map from the Internet
+  - way_filters['auto']
+  - add_reverse_edge: add reverse edge of two-way road. The order of reverse edge is [-1, -2, .., -n-1], where n in the number of line segments.
+  - A star 增加终止条件; Add `max_layer` parameter to control the deepest level.
+
+- PickleSaver
+  - Save the object to file, and load the object from the file.
+
+- setting.py
+  - Define `way_filters['auto']`
+
+### Changed
+
+- matching.py
+  - cal_relative_offset
+    Update the calculation process
+  - st_matching
+    plot with more details
+
 ## [1.0.01] - 2021-07-15
 
 ### Added